Refactor LibrisService to split ElasticSearch functionality into its own service
Fixes LibrisService is overweight #254
